+++ inst/doc/changing_funnel_plot_options.R
@@ -2,11 +2,11 @@
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
-  fig.height= 5, 
-  fig.width=7
+  fig.height = 5, 
+  fig.width = 7
 )
 
-## ----setup, fig.height= 5, fig.width=7----------------------------------------
+## ----setup--------------------------------------------------------------------
 library(FunnelPlotR)
 
 ## ----dtsetup------------------------------------------------------------------
@@ -18,20 +18,26 @@ medpar$los<-as.numeric(medpar$los)
 # Logistic model to predict LOS, LOS is quite overdispersed
 mod<- glm(los ~ hmo + died + age80 + factor(type), family="poisson", data=medpar)
 
-# Get predicted value for ratio
-medpar$prds<- predict(mod, type="response")
+#Get predicted value for ratio
+medpar$prds<- predict(mod, newdata = medpar, type="response")
 
 # Draw plot, returning just the plot object
-funnel_plot(denominator=medpar$prds,numerator=medpar$los
-            , group = medpar$provnum, limit=99 ,label_outliers = TRUE
-            , Poisson_limits = TRUE)
+funnel_plot(denominator=medpar$prds, numerator=medpar$los
+            , group = medpar$provnum, limit=99 ,label = "outlier"
+            , draw_unadjusted = TRUE)
 
+
+## ----highlight----------------------------------------------------------------
+# Draw plot, returning just the plot object
+funnel_plot(denominator=medpar$prds,numerator=medpar$los
+            , group = medpar$provnum, limit=99 ,label = "outlier"
+            , draw_unadjusted = TRUE, highlight="030002")
 
 ## ----plottheme1---------------------------------------------------------------
 
 funnel_plot(denominator=medpar$prds,numerator=medpar$los
-            , group = medpar$provnum, limit=99 ,label_outliers = TRUE
-            , Poisson_limits = TRUE, theme = funnel_grey() )
+            , group = medpar$provnum, limit=99 ,label = "outlier"
+            , draw_unadjusted = TRUE, theme = funnel_grey() )
 
 
 ## ----plottheme2---------------------------------------------------------------
@@ -43,37 +49,45 @@ new_funnel_theme <-
           legend.background = element_rect(fill="brown"), # Alter legend background colour
           axis.title.y = element_text(angle=0)  #Rotate y axis label
     )
-  
 
-funnel_plot(denominator=medpar$prds,numerator=medpar$los 
-            , group = medpar$provnum, limit=99 ,label_outliers = TRUE
-            , Poisson_limits = TRUE, theme = new_funnel_theme)
+
+funnel_plot(denominator=medpar$prds,numerator=medpar$los
+            , group = medpar$provnum, limit=99 ,label = "outlier"
+            , draw_unadjusted = TRUE, theme = new_funnel_theme)
 
 
 ## ----colours------------------------------------------------------------------
-funnel_plot(denominator=medpar$prds,numerator=medpar$los 
-            , group = medpar$provnum, limit=99 ,label_outliers = TRUE
-            , Poisson_limits = TRUE, plot_cols= c("#8c8c8c", "#00b159", "#00aedb", "#d11141"))
+funnel_plot(denominator=medpar$prds,numerator=medpar$los
+            , group = medpar$provnum, limit=99 ,label = "outlier"
+            , draw_unadjusted = TRUE, theme = funnel_grey(),
+            plot_cols = c("#FF7F0EFF", "#000000", "#1F77B4FF","#1F77B4FF", "#9467BDFF", "#9467BDFF", "#2CA02CFF", "#2CA02CFF"))
 
 ## ----funnelscales-------------------------------------------------------------
 ## Changing labels
-funnel_plot(denominator=medpar$prds,numerator=medpar$los 
-            , group = medpar$provnum, limit=99 ,label_outliers = TRUE
-            , Poisson_limits = TRUE, xrange=c(0, 400), yrange=c(0,2))
+funnel_plot(denominator=medpar$prds,numerator=medpar$los
+            , group = medpar$provnum, limit=99 ,label = "outlier"
+            , draw_unadjusted = TRUE, xrange=c(0, 400), yrange=c(0,2))
 
 
-## ----funnellabels-------------------------------------------------------------
-funnel_plot(denominator=medpar$prds,numerator=medpar$los 
-            , group = medpar$provnum, limit=99 ,label_outliers = TRUE
-            , Poisson_limits = TRUE, title = "Vignette funnel plot"
+## ----funnellabels1------------------------------------------------------------
+funnel_plot(denominator=medpar$prds,numerator=medpar$los
+            , group = medpar$provnum, limit=99 ,label = "outlier"
+            , draw_unadjusted = TRUE, title = "Vignette funnel plot"
             , x_label = "x-axis", y_label = "y-axis")
+
+## ----funnellabels2------------------------------------------------------------
+funnel_plot(denominator=medpar$prds,numerator=medpar$los
+            , group = medpar$provnum, limit=99
+            , draw_unadjusted = TRUE, title = "Vignette funnel plot"
+            , x_label = "x-axis", y_label = "y-axis"
+            , highlight= "030002", label = "highlight")
 
 ## ----cutoutplot---------------------------------------------------------------
 # Original funnel plot object
 fp <-
 funnel_plot(denominator=medpar$prds,numerator=medpar$los
-            , group = medpar$provnum, limit=99 ,label_outliers = TRUE
-            , Poisson_limits = TRUE)
+            , group = medpar$provnum, limit=99, label = "outlier"
+            , draw_unadjusted = TRUE)
 
 # Extract just the plot
 my_plot <- plot(fp)